Better heap implementation
With proper heap implementation, time complexity is O(nlogn), where n is the number of knots happen before the desired lambda. The key point is to accelerate how to modified a value in a heap. This is done by adding a vector that maps the id of beta's to the id of heaps. See the `map_to_heap` member in class `FusionGroups`